data-engine/requirements.txt
flask
psycopg2
pandas
requests
numpy
sentry-sdk
python-decouple
